fix:修复重构后的一些bug

+++ mooc-core/src/main/java/cn/edu/lingnan/mooc/core/vo/ChapterVO.java
@@ -17,11 +17,11 @@ public class ChapterVO {
     /**
      * 大章id
      */
-    private Integer id;
+    private Long id;
     /*
      * 对应课程id
      */
-    private Integer courseId;
+    private Long courseId;
     /*
      * 课程名称
      */
